Wire up thermostat configuration
This introduces the configuration core objects needed to offer
configuration options. I am still going to improve how they relate to
the representative views a bit more.

The thermostat checkbox is the only thing that works, but it does indeed
alter the total price now.

I tried to use the more "simple" approach of letting the view drive the
options but the substitution in the repetition made for some confusing
setup. Or I'm really incapable of letting go of this separation between
the data and the views.

As a nod to packaging convenience I did put the configuration inside
it's related panel .reel