Populate static view with some images
Just enough to get started; I haven't wired this up to react to changes
in the configuration. Frankly, that's a lot of tedious work to get
the right screenshots. I'm going to look into cleanly dividing
the 3d and static components outside of main; there's a lot of 3d
hardcoded into main right now.