Make static view react to configuration changes
If we come up with a way to show all of the options in here I'm game;
the system is flexible.

Anyway, now this should "work" on systems that don't have WebGL,
it's also possible to switch between the two views manually on systems
where you do have WebGL.

It might be nice to react to low battery state or something and switch
to the static view.

All images are placeholders for now.