Move configuration creation to applicationDelegate
This moves a lot of bootstrapping out of the component tree and into
the application code. Eventually that would be the place where
configuration elections could be loaded from a service or some data
storage.

This also moved the ordered panelDescriptors into the declaration for
easy of manipulation without having to drop into "code".

Furthermore, some names were changed and updated to better reflect
their role.

This is a stepping stone towards making sure the summary component
lists its groups in the same order as that the panels appear in,
without user intervention. Turns out they were simply done in order
as entered into the configuration, which is not ideal.

This is a fairly large change for what amounts to hopefully no
discernible difference; watch for regressions.