Fix nav dismissal pressComposer with touch
This updates montage to a quick fix such that a pressComposer correctly
ends interaction by removing the eventListeners it installed when the
interaction started.

More importantly it cancelsPress on the pressComposer after pressStart,
freeing other pressComposers further down the line to claim the now
surrendered pointer.

Also, there are some side-effects to the pressComposer being declared
in the template (it loads) which we now avoid by creating it
programmatically and unloading it early and often.